+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [core] Core will now throw an error when settings cannot be loaded
+- [core] Canary container images now run from source and include debug tools
 
 ### Fixed